+++ examples/web/src/discovery/index.ts
@@ -3,5 +3,6 @@ export * from './UseExploreProfiles';
 export * from './UseExplorePublications';
 export * from './UseFeed';
 export * from './UseFeedHighlights';
+export * from './UseRecommendedProfiles';
 export * from './UseSearchProfiles';
 export * from './UseSearchPublications';