+++ examples/web/src/profiles/index.ts
@@ -14,7 +14,6 @@ export * from './UseProfileFollowing';
 export * from './UseProfileManagers';
 export * from './UseProfiles';
 export * from './UseProfilesManaged';
-export * from './UseRecommendedProfiles';
 export * from './UseSetProfileMetadata';
 export * from './UseUpdateFollowPolicy';
 export * from './UseUpdateProfileManagers';