+++ packages/react/src/transactions/adapters/profiles/LinkHandleGateway.ts
@@ -1,14 +1,14 @@
 import {
-  SafeApolloClient,
-  omitTypename,
-  RelaySuccess,
-  CreateLinkHandleToProfileTypedDataVariables,
-  LinkHandleToProfileVariables,
-  LinkHandleToProfileDocument,
-  LinkHandleToProfileData,
+  CreateLinkHandleToProfileBroadcastItemResult,
   CreateLinkHandleToProfileTypedDataData,
   CreateLinkHandleToProfileTypedDataDocument,
-  CreateLinkHandleToProfileBroadcastItemResult,
+  CreateLinkHandleToProfileTypedDataVariables,
+  LinkHandleToProfileData,
+  LinkHandleToProfileDocument,
+  LinkHandleToProfileVariables,
+  RelaySuccess,
+  SafeApolloClient,
+  omitTypename,
 } from '@lens-protocol/api-bindings';
 import { lensTokenHandleRegistry } from '@lens-protocol/blockchain-bindings';
 import { NativeTransaction, Nonce } from '@lens-protocol/domain/entities';
@@ -22,19 +22,24 @@ import { ChainType, Data, PromiseResult, success } from '@lens-protocol/shared-k
 import { v4 } from 'uuid';
 
 import { UnsignedProtocolCall } from '../../../wallet/adapters/ConcreteWallet';
+import { IProviderFactory } from '../../../wallet/adapters/IProviderFactory';
+import { AbstractContractCallGateway, ContractCallDetails } from '../AbstractContractCallGateway';
 import { ITransactionFactory } from '../ITransactionFactory';
-import { SelfFundedProtocolTransactionRequest } from '../SelfFundedProtocolTransactionRequest';
 import { handleRelayError } from '../relayer';
 
 export class LinkHandleGateway
+  extends AbstractContractCallGateway<LinkHandleRequest>
   implements
     IDelegatedTransactionGateway<LinkHandleRequest>,
     ISignedOnChainGateway<LinkHandleRequest>
 {
   constructor(
+    providerFactory: IProviderFactory,
     private readonly apolloClient: SafeApolloClient,
     private readonly transactionFactory: ITransactionFactory<LinkHandleRequest>,
-  ) {}
+  ) {
+    super(providerFactory);
+  }
 
   async createDelegatedTransaction(
     request: LinkHandleRequest,
@@ -64,8 +69,12 @@ export class LinkHandleGateway
       id: result.id,
       request,
       typedData: omitTypename(result.typedData),
-      fallback: this.createRequestFallback(request, result),
     });
+  }
+
+  protected async createEncodedData(request: LinkHandleRequest): Promise<ContractCallDetails> {
+    const result = await this.createTypedData(request);
+    return this.createLinkCallDetails(result);
   }
 
   private async relayWithProfileManager(
@@ -84,10 +93,7 @@ export class LinkHandleGateway
     });
 
     if (data.result.__typename === 'LensProfileManagerRelayError') {
-      const result = await this.createTypedData(request);
-      const fallback = this.createRequestFallback(request, result);
-
-      return handleRelayError(data.result, fallback);
+      return handleRelayError(data.result);
     }
 
     return success(data.result);
@@ -110,17 +116,15 @@ export class LinkHandleGateway
     return data.result;
   }
 
-  private createRequestFallback(
-    request: LinkHandleRequest,
+  private createLinkCallDetails(
     result: CreateLinkHandleToProfileBroadcastItemResult,
-  ): SelfFundedProtocolTransactionRequest<LinkHandleRequest> {
+  ): ContractCallDetails {
     const contract = lensTokenHandleRegistry(result.typedData.domain.verifyingContract);
     const encodedData = contract.interface.encodeFunctionData('link', [
       result.typedData.message.handleId,
       result.typedData.message.profileId,
     ]);
     return {
-      ...request,
       contractAddress: result.typedData.domain.verifyingContract,
       encodedData: encodedData as Data,
     };