+++ packages/react/src/transactions/adapters/profiles/__tests__/LinkHandleGateway.spec.ts
@@ -1,3 +1,6 @@
+/*
+ * @jest-environment node
+ */
 import { SafeApolloClient } from '@lens-protocol/api-bindings';
 import {
   mockLensApolloClient,
@@ -6,24 +9,63 @@ import {
   mockCreateLinkHandleToProfileTypedDataResponse,
   mockLinkHandleToProfileResponse,
 } from '@lens-protocol/api-bindings/mocks';
-import { NativeTransaction } from '@lens-protocol/domain/entities';
-import { mockLinkHandleRequest } from '@lens-protocol/domain/mocks';
+import { NativeTransaction, UnsignedTransaction } from '@lens-protocol/domain/entities';
+import { mockLinkHandleRequest, mockWallet } from '@lens-protocol/domain/mocks';
+import { ChainType } from '@lens-protocol/shared-kernel';
+import { providers } from 'ethers';
+import { mock } from 'jest-mock-extended';
 
 import { UnsignedProtocolCall } from '../../../../wallet/adapters/ConcreteWallet';
+import { mockIProviderFactory } from '../../../../wallet/adapters/__helpers__/mocks';
+import { UnsignedContractCallTransaction } from '../../AbstractContractCallGateway';
 import { assertUnsignedProtocolCallCorrectness } from '../../__helpers__/assertions';
-import { mockITransactionFactory } from '../../__helpers__/mocks';
+import { mockITransactionFactory, mockJsonRpcProvider } from '../../__helpers__/mocks';
 import { LinkHandleGateway } from '../LinkHandleGateway';
 
-function setupTestScenario({ apolloClient }: { apolloClient: SafeApolloClient }) {
+function setupTestScenario({
+  apolloClient,
+  provider = mock<providers.JsonRpcProvider>(),
+}: {
+  apolloClient: SafeApolloClient;
+  provider?: providers.JsonRpcProvider;
+}) {
   const transactionFactory = mockITransactionFactory();
+  const providerFactory = mockIProviderFactory({
+    chainType: ChainType.POLYGON,
+    provider,
+  });
 
-  const gateway = new LinkHandleGateway(apolloClient, transactionFactory);
+  const gateway = new LinkHandleGateway(providerFactory, apolloClient, transactionFactory);
 
   return { gateway };
 }
 
 describe(`Given an instance of ${LinkHandleGateway.name}`, () => {
   const request = mockLinkHandleRequest();
+
+  describe(`when creating an ${UnsignedTransaction.name}<LinkHandleRequest>`, () => {
+    const wallet = mockWallet();
+    const data = mockCreateLinkHandleToProfileTypedDataData();
+
+    it(`should succeed with the expected ${UnsignedContractCallTransaction.name}`, async () => {
+      const provider = await mockJsonRpcProvider();
+      const apolloClient = mockLensApolloClient([
+        mockCreateLinkHandleToProfileTypedDataResponse({
+          variables: {
+            request: {
+              handle: request.fullHandle,
+            },
+          },
+          data,
+        }),
+      ]);
+      const { gateway } = setupTestScenario({ apolloClient, provider });
+
+      const unsignedTransaction = await gateway.createUnsignedTransaction(request, wallet);
+
+      expect(unsignedTransaction).toBeInstanceOf(UnsignedContractCallTransaction);
+    });
+  });
 
   describe(`when creating an IUnsignedProtocolCall<LinkHandleRequest>`, () => {
     it(`should create an instance of the ${UnsignedProtocolCall.name} with the expected typed data`, async () => {