+++ packages/react/src/transactions/useBlockProfiles.ts
@@ -151,7 +151,7 @@ export type BlockOperation = UseDeferredTask<
  *     profiles: [profileId],
  *   });
  *
- * if (result.isFailure()) {
+ *   if (result.isFailure()) {
  *     // handle failure scenarios
  *     return;
  *   }
@@ -216,7 +216,6 @@ export type BlockOperation = UseDeferredTask<
  *   // sponsored attempt
  *   const sponsoredResult = await execute({
  *     profiles: [profileId],
- *     sponsored: false
  *   });
  *
  *   if (sponsoredResult.isFailure()) {
@@ -225,8 +224,8 @@ export type BlockOperation = UseDeferredTask<
  *         if ([BroadcastingErrorReason.NOT_SPONSORED, BroadcastingErrorReason.RATE_LIMITED].includes(sponsoredResult.error.reason)) {
  *
  *           const selfFundedResult = await execute({
- *             metadataURI: uri,
- *             sponsored: false
+ *              profiles: [profileId],
+ *              sponsored: false
  *           });
  *
  *           // continue with selfFundedResult as in the previous example
@@ -259,7 +258,7 @@ export function useBlockProfiles(): BlockOperation {
       profileIds: profiles.map((profile) => profile.id),
       kind: TransactionKind.BLOCK_PROFILE,
       signless: session.profile.signless,
-      sponsored: sponsored ?? false,
+      sponsored: sponsored ?? true,
     });
   });
 }