+++ packages/react/src/transactions/useCreateComment.ts
@@ -41,7 +41,7 @@ export type CreateCommentArgs = {
   /**
    * The Open Actions associated with the publication.
    *
-   * If none provided the comment will be automatically hosted on Momoka.
+   * @defaultValue empty, no open actions
    */
   actions?: OpenActionConfig[];
   /**
@@ -270,10 +270,6 @@ export type CreateCommentArgs = {
  *
  * Contextually to the comment creation you can configure the reference policy.
  *
- * A comment with reference policy other than `ANYONE` will be hosted on-chain.
- * If the comment has reference policy `ANYONE` (which is also the default value) and does not have
- * any open actions, it will be hosted on Momoka.
- *
  * No one can comment, quote, or mirror the comment:
  * ```tsx
  * const result = await execute({
@@ -358,7 +354,7 @@ export type CreateCommentArgs = {
  * });
  * ```
  *
- * ## Self-funded Comment
+ * ## Self-funded approach
  *
  * In case you want to pay for the transaction gas costs yourself, you can do so by setting the
  * `sponsored` parameter to `false`:
@@ -390,7 +386,7 @@ export type CreateCommentArgs = {
  *
  * The example above shows how to detect when the user does not have enough funds to pay for the transaction cost.
  *
- * ## Self-funded Fallback
+ * ## Self-funded fallback
  *
  * If for some reason the Lens API cannot sponsor the transaction, the hook will fail with a {@link BroadcastingError} with one of the following reasons:
  * - {@link BroadcastingErrorReason.NOT_SPONSORED} - the profile is not sponsored
@@ -433,6 +429,15 @@ export type CreateCommentArgs = {
  *
  * It just requires the app to apply for whitelisting. See https://docs.lens.xyz/docs/gasless-and-signless#whitelisting-your-app.
  *
+ * ## Momoka comments
+ *
+ * For a comment to be hosted on Momoka it must meet the following criteria:
+ * - it must be a comment for a Momoka publication
+ * - reference policy `ANYONE` (which is also the default value in case it's not specified)
+ * - no open actions
+ * - sponsored by the Lens API (which is also the default value in case it's not specified)
+ *
+ * If the comment does not meet the above criteria, it will be hosted on-chain.
  *
  * ## Upgrading from v1
  *