+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @lens-protocol/react-native
 
+## 0.1.0-alpha.2
+
+### Minor Changes
+
+- d255b3627: **feat:** Added `useDismissRecommendedProfiles` hook
+
+### Patch Changes
+
+- daf688200: **fixed:** logic that infers when to create Momoka publications
+- Updated dependencies [daf688200]
+- Updated dependencies [d255b3627]
+- Updated dependencies [dd5088811]
+- Updated dependencies [c6da5071d]
+- Updated dependencies [b8279c3bd]
+  - @lens-protocol/react@2.0.0-alpha.25
+  - @lens-protocol/domain@0.11.0-alpha.17
+  - @lens-protocol/api-bindings@0.11.0-alpha.22
+
 ## 0.0.1-alpha.1
 
 ### Patch Changes